List now shows apartment objects
The list now shows the apartment objects and the markers now show up on the google map. The only thing is you have to manually type the command initMap() in the console. If you run it in the code itself, the site gets hung. Next thing that needs to be fixed.